# postgres-project-tenancy/openapi.yaml
openapi: 3.1.0
info:
  title: Project Tenancy API
  version: 1.0.0
  description: |
    Project-scoped multi-tenant API. Project membership is the single shared
    access level. Owners retain only lifecycle and sharing administration;
    owners and shared members have identical service/resource CRUD access.
servers:
  - url: https://api.example.com/v1
    description: Replace with the deployed API origin
security:
  - oidc:
      - openid
      - profile
      - email
...
paths:
...
  /projects:
    get:
      tags: [Projects]
      operationId: listProjects
      summary: List all projects visible to the current user
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: ownership
          in: query
          description: Optionally restrict results to owned or shared projects.
          schema:
            type: string
            enum: [all, owned, shared]
            default: all
      responses:
        '200':
          description: Page of projects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Projects]
      operationId: createProject
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created; owner membership is created atomically.
          headers:
            Location:
              $ref: '#/components/headers/Location'
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /projects/{projectId}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [Projects]
      operationId: getProject
      summary: Get a visible project
      responses:
        '200':
          description: Project
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Projects]
      operationId: updateProject
      summary: Update project metadata
      description: Owner only.
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: Updated project
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags: [Projects]
      operationId: deleteProject
      summary: Permanently delete a project and all descendants
      description: Owner only. A production UI should require explicit confirmation.
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '204':
          description: Project deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
...
  /projects/{projectId}/services/{serviceId}/resources/{resourceId}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/ServiceId'
      - $ref: '#/components/parameters/ResourceId'
    get:
      tags: [Resources]
      operationId: getServiceResource
      summary: Get a service resource
      responses:
        '200':
          description: Resource
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Resources]
      operationId: updateServiceResource
      summary: Update a service resource
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/UpdateResourceRequest'
      responses:
        '200':
          description: Updated resource
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags: [Resources]
      operationId: deleteServiceResource
      summary: Delete a service resource
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '204':
          description: Resource deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      description: |
        Authorization Code with PKCE. Replace the URL with the configured
        provider. The API validates access tokens for its own audience.
      openIdConnectUrl: https://identity.example.com/.well-known/openid-configuration
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ServiceId:
      name: serviceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ResourceId:
      name: resourceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Limit:
      name: limit
      in: query
      description: Maximum number of results.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor returned by the previous page.
      schema:
        type: string
        maxLength: 2000
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Quoted integer version from the latest ETag, for example `"7"`.
      schema:
        type: string
        pattern: '^"[1-9][0-9]*"$'
  headers:
    ETag:
      description: Quoted value of the resource's integer version.
      schema:
        type: string
        pattern: '^"[1-9][0-9]*"$'
    Location:
      description: Canonical URL of the created resource.
      schema:
        type: string
        format: uri-reference
  responses:
    Unauthorized:
      description: Missing, invalid, expired, or wrong-audience access token.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: The authenticated user cannot perform this operation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: The resource does not exist or is outside the caller's visibility.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Unique constraint or resource-state conflict.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    PreconditionFailed:
      description: If-Match does not match the current resource version.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ValidationError:
      description: Request syntax is valid but one or more values are invalid.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    Project:
      type: object
      additionalProperties: false
      required:
        - id
        - owner_user_id
        - name
        - access_type
        - version
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        owner_user_id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: [string, 'null']
          maxLength: 5000
        access_type:
          type: string
          description: Ownership marker for UI/admin decisions, not a membership role.
          enum: [owner, shared]
        version:
          type: integer
          format: int64
          minimum: 1
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateProjectRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: [string, 'null']
          maxLength: 5000
    UpdateProjectRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: [string, 'null']
          maxLength: 5000
...
    Resource:
      type: object
      additionalProperties: false
      required:
        - project_id
        - service_id
        - id
        - resource_key
        - payload
        - version
        - created_at
        - updated_at
      properties:
        project_id:
          type: string
          format: uuid
        service_id:
          type: string
          format: uuid
        id:
          type: string
          format: uuid
        resource_key:
          type: string
          minLength: 1
          maxLength: 200
        payload:
          type: object
          additionalProperties: true
        version:
          type: integer
          format: int64
          minimum: 1
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    UpdateResourceRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        resource_key:
          type: string
          minLength: 1
          maxLength: 200
        payload:
          type: object
          additionalProperties: true
...
    ProjectPage:
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Project'
        next_cursor:
          type: [string, 'null']
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
          example: https://api.example.com/problems/validation-error
        title:
          type: string
          example: Request validation failed
        status:
          type: integer
          minimum: 400
          maximum: 599
          example: 422
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
          example: validation_error
        request_id:
          type: string
          format: uuid
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
              message:
                type: string
      additionalProperties: true
...